client/daemon/pyproject.toml
[project]
name = "racksmith-daemon"
version = "0.1.0"
description = "Racksmith daemon — SSH, Ansible, and network operations on the user's LAN"
requires-python = ">=3.14"
license = "AGPL-3.0"
authors = [{name = "Ekku Leivonen"}]

dependencies = [
    "ansible-core>=2.20.3",
    "arq>=0.25.0",
    "asyncssh>=2.22.0",
    "fastapi[standard]>=0.135.1",
    "mac-vendor-lookup>=0.1.15",
    "pyyaml>=6.0.3",
    "racksmith-shared",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.uv.sources]
racksmith-shared = { path = "../shared", editable = true }

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true
strict_optional = true
